Revert "SWDEV-505971 - change setArgument arg from uint32_t to uint64_t"
This reverts commit a6f2a2c2afa388e69bc7389622dfec480a2f35bb.

Reason for revert: There needs to be memcpy size change

+++ rocclr/device/rocm/rocblit.cpp
@@ -2513,7 +2513,7 @@ bool KernelBlitManager::streamOpsWait(device::Memory& memory, uint64_t value, si
 
 // ================================================================================================
 bool KernelBlitManager::batchMemOps(const void* paramArray, size_t paramSize,
-                                    uint64_t count) const {
+                                    uint32_t count) const {
   amd::ScopedLock k(lockXferOps_);
   bool result = false;
   uint blitType = BatchMemOp;
